+++ transtats/dashboard/managers/packages.py
@@ -13,7 +13,10 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 
+from datetime import datetime
+
 from ..models.package import Packages
+from ..models.transplatform import TransPlatform
 from .base import BaseManager
 
 
@@ -26,10 +29,42 @@ class PackagesManager(BaseManager):
         """
         fetch packages from db
         """
+        packages = None
         try:
             packages = self.db_session.query(Packages).order_by('transtats_lastupdated').all()
         except:
             # log event, passing for now
             pass
+        return packages
+
+    def add_package(self, **kwargs):
+        """
+        add package to db
+        :param kwargs: dict
+        :return: boolean
+        """
+        required_params = ('package_name', 'upstream_url', 'transplatform_slug', 'release_stream_slug')
+        if not set(required_params) < set(kwargs.keys()):
+            return
+
+        if not (kwargs.get('package_name') and kwargs.get('upstream_url')):
+            return
+
+        try:
+            # derive transplatform project URL
+            platform_url = self.db_session.query(TransPlatform.api_url). \
+                filter_by(platform_slug=kwargs['transplatform_slug']).one()[0]
+            kwargs['transplatform_url'] = platform_url + "/project/view/" + kwargs['package_name']
+            # override lang_set and transtats_lastupdated values
+            kwargs['lang_set'] = 'default'
+            kwargs['transtats_lastupdated'] = datetime.now()
+            # save in db
+            new_package = Packages(**kwargs)
+            self.db_session.add(new_package)
+            self.db_session.commit()
+        except:
+            self.db_session.rollback()
+            # log event, pass for now
+            return False
         else:
-            return packages
+            return True